rename v2_compat -> _compat
