+++ tests/experimental/pydantic/schema/test_defaults.py
@@ -5,7 +5,7 @@ import pydantic
 
 import strawberry
 from strawberry.printer import print_schema
-from tests.experimental.pydantic.utils import needs_pydanticv2
+from tests.experimental.pydantic.utils import needs_pydantic_v2
 
 
 def test_field_type_default():
@@ -159,7 +159,7 @@ def test_input_type_default():
     assert print_schema(schema) == textwrap.dedent(expected).strip()
 
 
-@needs_pydanticv2
+@needs_pydantic_v2
 def test_v2_explicit_default():
     class User(pydantic.BaseModel):
         name: Optional[str]
